New version 301: various new effects, changes to sienaRI, some smaller changes.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rsiena@301 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/RSienaTest/src/data/NetworkLongitudinalData.h
@@ -111,7 +111,7 @@ private:
 	int lmaxDegree;
 
 	// The offset for the non-primary settings. 0 by default.
-	int luniversalOffset;
+	double luniversalOffset;
 
 	// The average in-degree over all receivers and observations
 	double laverageInDegree;
